Make bug fixes
