+++ backend/src/xfd_django/xfd_api/tasks/refresh_vs_summaries.py
@@ -1,12 +1,16 @@
 """Run Summary population methods via a scan."""  # Standard Python Libraries
 # Standard Python Libraries
+from datetime import timedelta
 import logging
 import os
+import random
+
+# Third-Party Libraries
+from django.utils import timezone
 
 logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
 LOGGER = logging.getLogger(__name__)
 # Third-Party Libraries
-from xfd_api.tasks.syncdb_task import build_fake_host_summaries
 from xfd_api.tasks.vulnScanningSync import (
     create_daily_host_summary,
     create_port_scan_service_summaries,
@@ -14,7 +18,7 @@ from xfd_api.tasks.vulnScanningSync import (
     create_vuln_scan_summary,
     enforce_latest_flag_port_scan,
 )
-from xfd_mini_dl.models import Organization
+from xfd_mini_dl.models import HostSummary, Organization
 
 
 def rebuild_org_id_dict(db_name="mini_data_lake"):
@@ -24,6 +28,52 @@ def rebuild_org_id_dict(db_name="mini_data_lake"):
         for org in Organization.objects.using(db_name).all()
         if org.acronym  # defensive check
     }
+
+
+def build_fake_host_summaries():
+    """Build a fake Ticket for a pssed org."""
+    all_orgs = Organization.objects.all()
+
+    for org in all_orgs:
+        try:
+            summary_date = timezone.now().date()
+            start_date = timezone.now() - timedelta(
+                days=random.randint(25, 60), seconds=random.randint(0, 86400)
+            )
+            end_date = timezone.now() - timedelta(
+                days=random.randint(1, 5), seconds=random.randint(0, 86400)
+            )
+            host_done_count = random.randint(3000, 5000)
+            host_waiting_count = random.randint(0, 50)
+            host_running_count = random.randint(0, 50)
+            host_ready_count = random.randint(0, 50)
+            total_count = (
+                host_done_count
+                + host_waiting_count
+                + host_running_count
+                + host_ready_count
+            )
+            up_host_count = total_count - random.randint(0, 1500)
+            down_host_count = total_count - up_host_count
+
+            HostSummary.objects.update_or_create(
+                organization=org,
+                summary_date=summary_date,
+                defaults={
+                    "start_date": start_date,
+                    "end_date": end_date,
+                    "host_done_count": host_done_count,
+                    "host_waiting_count": host_waiting_count,
+                    "host_running_count": host_running_count,
+                    "host_ready_count": host_ready_count,
+                    "up_host_count": up_host_count,
+                    "down_host_count": down_host_count,
+                    "scanned_asset_count": total_count,
+                },
+            )
+        except Exception as e:
+            print("\n❌ Error while creating host_summary for org %s: %s", org.name, e)
+            continue
 
 
 def handler(event):
